Change chart implementation to React Hooks and impelemented caching for API call
