+++ src/components/PaperTrade.js
@@ -1,0 +1,10 @@
+import React, { useState, useEffect } from "react";
+import Typography from "@material-ui/core/Typography";
+
+export default function PaperTrade() {
+  return (
+    <div className="papertrade">
+      <Typography variant="h2" color="primary">PAPER TRADE</Typography>
+    </div>
+  );
+}
