+++ src/components/Watchlist.js
@@ -1,0 +1,10 @@
+import React, { useState, useEffect } from "react";
+import Typography from "@material-ui/core/Typography";
+
+export default function Watchlist() {
+  return (
+    <div className="watchlist">
+      <Typography variant="h2" color="primary">MY WATCHLIST</Typography>
+    </div>
+  );
+}
