+++ src/styles/CompanyOverview.css
@@ -1,0 +1,7 @@
+.symbol {
+  height: 36px;
+}
+
+.industry {
+  height: 50px;
+}
